+++ src/main/resources/static/payment.html
@@ -23,7 +23,7 @@
           IMP.init("imp51203488"); // PG사 사업자 IMP번호 (Secret으로 관리되어야 함)
 
           //임시로 넣어놓은 데이터
-          var orderuid = "7973f635-c3a4-4535-a5af-54be62c3fd58"
+          var orderuid = "2ab40bbe-fe4d-4122-9725-dd8ed6d1deab"
           var item_name = "premium"
           var buyer_email = "asdf1234@naver.com"
           var buyer_name = "홍길동"
